Pull out UnableToCreatePlayerException.
